Refactor: Gamma encoding via CameraSettings instead of DisplayConfiguration
DisplayConfiguration contains values that depend on the display and are independent of all cameras.
CameraSettings contains values that depend on the camera and can thus vary between cameras.
Unfortunately, due to practical design decisions it is not feasable (e.g., performance considerations) to allow variation of each such value between cameras.